Centralise log references
