+++ api/consumer/searchDocumentReference/src/v1/handler.py
@@ -1,4 +1,3 @@
-from enum import Enum
 from logging import Logger
 from typing import Any
 
@@ -9,12 +8,9 @@ from nrlf.core.common_steps import make_common_log_action, parse_headers
 from nrlf.core.model import APIGatewayProxyEventModel, ConsumerRequestParams
 from nrlf.core.repository import Repository
 from nrlf.core.transform import create_bundle_from_paginated_response
+from nrlf.log_references import LogReference
 
 log_action = make_common_log_action()
-
-
-class LogReference(Enum):
-    SEARCH001 = "Searching for document references"
 
 
 @log_action(log_reference=LogReference.SEARCH001)
