+++ api/consumer/searchPostDocumentReference/src/v1/handler.py
@@ -1,4 +1,3 @@
-from enum import Enum
 from logging import Logger
 from typing import Any
 
@@ -10,12 +9,9 @@ from nrlf.core.event_parsing import fetch_body_from_event
 from nrlf.core.model import APIGatewayProxyEventModel, ConsumerRequestParams
 from nrlf.core.repository import Repository
 from nrlf.core.transform import create_bundle_from_paginated_response
+from nrlf.log_references import LogReference
 
 log_action = make_common_log_action()
-
-
-class LogReference(Enum):
-    SEARCHPOST001 = "Searching for document references"
 
 
 @log_action(log_reference=LogReference.SEARCHPOST001)
