+++ firehose/alert/steps.py
@@ -3,8 +3,9 @@ import gzip
 import requests
 from lambda_utils.logging import log_action
 
-from firehose.alert.constants import DUMMY_URL, FLAKY_SPLUNK_ERROR_CODES, LogReference
+from firehose.alert.constants import DUMMY_URL, FLAKY_SPLUNK_ERROR_CODES
 from nrlf.core.validators import json_loads
+from nrlf.log_references import LogReference
 
 
 @log_action(
