+++ layer/nrlf/nrlf/core/firehose/utils.py
@@ -8,8 +8,8 @@ from aws_lambda_powertools.utilities.parser.models.kinesis_firehose import (
 )
 from lambda_utils.logging import log_action
 
-from nrlf.core.firehose.log_reference import LogReference
 from nrlf.core.validators import json_loads
+from nrlf.log_references import LogReference
 
 LOG_SEPARATOR = ""
